Closing #134 Using log dir argument for l2logger in rl_experiment() and run()
