+++ tests/test_rl_cli.py
@@ -11,10 +11,11 @@ from .simple_agent import SimpleRLAgent
 @patch(
     "argparse.ArgumentParser.parse_args",
     return_value=argparse.Namespace(
-        num_parallel_envs=1, num_lifetimes=1, log_dir="", render=False
+        num_parallel_envs=1, num_lifetimes=1, log_dir="logs", render=False
     ),
 )
-def test_no_args(p):
+def test_no_args(p, tmpdir):
+    tmpdir.chdir()
     # TODO what should this test other than being runnable?
     rl_cli(SimpleRLAgent, SimpleRLCurriculum)
 
@@ -22,10 +23,11 @@ def test_no_args(p):
 @patch(
     "argparse.ArgumentParser.parse_args",
     return_value=argparse.Namespace(
-        num_parallel_envs=1, num_lifetimes=2, log_dir="", render=False
+        num_parallel_envs=1, num_lifetimes=2, log_dir="logs", render=False
     ),
 )
-def test_num_lifetimes(p):
+def test_num_lifetimes(p, tmpdir):
+    tmpdir.chdir()
     # TODO what should this test other than being runnable?
     rl_cli(SimpleRLAgent, SimpleRLCurriculum)
 
@@ -33,10 +35,11 @@ def test_num_lifetimes(p):
 @patch(
     "argparse.ArgumentParser.parse_args",
     return_value=argparse.Namespace(
-        num_parallel_envs=2, num_lifetimes=1, log_dir="", render=False
+        num_parallel_envs=2, num_lifetimes=1, log_dir="logs", render=False
     ),
 )
-def test_num_parallel_envs(p):
+def test_num_parallel_envs(p, tmpdir):
+    tmpdir.chdir()
     # TODO what should this test other than being runnable?
     rl_cli(SimpleRLAgent, SimpleRLCurriculum)
 
@@ -46,12 +49,13 @@ def test_num_parallel_envs(p):
     return_value=argparse.Namespace(
         num_parallel_envs=1,
         num_lifetimes=2,
-        log_dir="",
+        log_dir="logs",
         curriculum="invalid",
         render=False,
     ),
 )
-def test_invalid_curriculum_name(p):
+def test_invalid_curriculum_name(p, tmpdir):
+    tmpdir.chdir()
     with pytest.raises(RuntimeError):
         rl_cli(SimpleRLAgent)
 
@@ -59,11 +63,12 @@ def test_invalid_curriculum_name(p):
 @patch(
     "argparse.ArgumentParser.parse_args",
     return_value=argparse.Namespace(
-        num_parallel_envs=1, num_lifetimes=2, log_dir="", render=False
+        num_parallel_envs=1, num_lifetimes=2, log_dir="logs", render=False
     ),
 )
 @patch("tella.env.L2LoggerEnv.render")
-def test_no_render(render_patch, argparse_patch):
+def test_no_render(render_patch, argparse_patch, tmpdir):
+    tmpdir.chdir()
     rl_cli(SimpleRLAgent, SimpleRLCurriculum)
     assert not render_patch.called
 
@@ -71,10 +76,11 @@ def test_no_render(render_patch, argparse_patch):
 @patch(
     "argparse.ArgumentParser.parse_args",
     return_value=argparse.Namespace(
-        num_parallel_envs=1, num_lifetimes=2, log_dir="", render=True
+        num_parallel_envs=1, num_lifetimes=2, log_dir="logs", render=True
     ),
 )
 @patch("tella.env.L2LoggerEnv.render")
-def test_renders(render_patch, argparse_patch):
+def test_renders(render_patch, argparse_patch, tmpdir):
+    tmpdir.chdir()
     rl_cli(SimpleRLAgent, SimpleRLCurriculum)
     assert render_patch.called
